use default comparer

+++ DataStructures/ExtensionMethods/Student.cs
@@ -8,9 +8,9 @@ namespace ExtensionMethods
     {
         public int ID { get; set; }
         public string Name { get; set; }
-        public int Age { get; set; }
+        public double Age { get; set; }
 
-        public Student (int id, string name, int age)
+        public Student (int id, string name, double age)
         {
             this.ID = id;
             this.Name = name;
